Fixed DateFormat reentrant bug.  Thanks to Bob Lacatena.
